src: Move Mcap/Channel to its own file
